Missing vendors, client test packages

+++ routes/server.go
@@ -18,7 +18,7 @@ type HTTPServer interface {
 }
 
 type server struct {
-	port      uint
+	port      uint16
 	accessLog io.Writer
 	projects  controller.Projects
 	router    *mux.Router
@@ -39,7 +39,7 @@ func projectsController() controller.Projects {
 }
 
 // InitializeServer creates an HTTPServer
-func InitializeServer(port uint, accessLog io.Writer, system controller.System, projects controller.Projects, siteDir string) HTTPServer {
+func InitializeServer(port uint16, accessLog io.Writer, system controller.System, projects controller.Projects, siteDir string) HTTPServer {
 	serverSingleton.init.Do(func() {
 		serverSingleton.instance = &server{
 			port:      port,
